pyproject.toml
[project]
name = "disignathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "bcrypt==4.0.1",
    "cryptography>=45.0.4",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.13",
    "itsdangerous>=2.2.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.24",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.5.0",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
